chore: reorganize examples into topic-grouped experiments tree
- Move examples/ to experiments/ with subdirs grouping by study
  (basic, cross_validation, lbfgsb, handoff, report)
- Rename scripts for clarity once grouped (e.g. lbfgsb_benchmark.py
  -> lbfgsb/sphere_benchmark.py, cmaes_handoff_study.py
  -> handoff/covariance_transformations.py)
- Update internal Path("plots/...") references inside the moved
  scripts so each experiment writes into plots/<group>/<name>/,
  mirroring the experiments tree
- Move root orphans into reference/ (R impls + their CSV outputs)
  and docs/ (plot_descriptions.md -> cmaes_diagnostic_plots.md)
- Update pyproject.toml entry points (run-example, run-r) and the
  package description
