+++ docs/cmaes_diagnostic_plots.md
